Fixes regarding feedback on PR
In resources changed using models to using resources instead. Moved following projects from user controller to project controller. Added comments and got rid of all warnings.
