+++ API/Controllers/UserController.cs
@@ -42,22 +42,21 @@ namespace API.Controllers
         private readonly IUserService userService;
         private readonly IRoleService roleService;
         private readonly IProjectService projectService;
-        private readonly IUserProjectService userProjectService;
         private readonly IUserUserService userUserService;
-
         /// <summary>
-        /// Initializes a new instance of the <see cref="UserController"/> class.
+        ///  Initializes a new instance of the <see cref="UserController"/> class.
         /// </summary>
-        /// <param name="userService">The user service.</param>
-        /// <param name="mapper">The mapper.</param>
-        /// <param name="roleService">The role service.</param>
-        public UserController(IUserService userService, IMapper mapper, IRoleService roleService, IUserProjectService userProjectService, IProjectService projectService,IUserUserService userUserService)
+        /// <param name="userService"></param>
+        /// <param name="mapper"></param>
+        /// <param name="roleService"></param>
+        /// <param name="projectService"></param>
+        /// <param name="userUserService"></param>
+        public UserController(IUserService userService, IMapper mapper, IRoleService roleService, IProjectService projectService,IUserUserService userUserService)
         {
             this.userService = userService;
             this.mapper = mapper;
             this.roleService = roleService;
             this.projectService = projectService;
-            this.userProjectService = userProjectService;
             this.userUserService = userUserService;
         }
 
@@ -222,109 +221,7 @@ namespace API.Controllers
             return Ok();
         }
 
-        /// <summary>
-        /// Follows a project with given projectId and gets userId
-        /// </summary>
-        /// <param name="projectId"></param>
-        /// <returns>200 if success 409 if user already follows project</returns>
-        [HttpPost("follow/project/{projectId}")]
-        [Authorize]
-        public async Task<IActionResult> FollowProject(int projectId)
-        {
-            User user = await HttpContext.GetContextUser(userService).ConfigureAwait(false);
-
-            if(await userService.FindAsync(user.Id) == null)
-            {
-                ProblemDetails problem = new ProblemDetails
-                {
-                    Title = "Failed getting the user account.",
-                    Detail = "The database does not contain a user with this user id.",
-                    Instance = "B778C55A-D41E-4101-A7A0-F02F76E5A6AE"
-                };
-                return NotFound(problem);
-            }
-
-            if(userProjectService.CheckIfUserFollows(user.Id,projectId))
-            {
-                ProblemDetails problem = new ProblemDetails
-                {
-                    Title = "User already follows this project",
-                    Detail = "You are already following this project.",
-                    Instance = "27D14082-9906-4EB8-AE4C-65BAEC0BB4FD"
-                };
-                return Conflict(problem);
-            }
-
-            Project project = await projectService.FindAsync(projectId);
-
-            if(await projectService.FindAsync(projectId) == null)
-            {
-                ProblemDetails problem = new ProblemDetails
-                {
-                    Title = "Failed getting the project.",
-                    Detail = "The database does not contain a project with this project id.",
-                    Instance = "57C13F73-6D22-41F3-AB05-0CCC1B3C8328"
-                };
-                return NotFound(problem);
-            }
-            UserProject userProject = new UserProject(project, user);
-            userProjectService.Add(userProject);
-
-            userProjectService.Save();
-            return Ok();
-        }
-
-        /// <summary>
-        /// Unfollows project
-        /// </summary>
-        /// <param name="projectId"></param>
-        /// <returns></returns>
-        [HttpDelete("follow/project/{projectId}")]
-        [Authorize]
-        public async Task<IActionResult> UnfollowProject(int projectId)
-        {
-            User user = await HttpContext.GetContextUser(userService).ConfigureAwait(false);
-
-            if(await userService.FindAsync(user.Id) == null)
-            {
-                ProblemDetails problem = new ProblemDetails
-                {
-                    Title = "Failed getting the user account.",
-                    Detail = "The database does not contain a user with this user id.",
-                    Instance = "B778C55A-D41E-4101-A7A0-F02F76E5A6AE"
-                };
-                return NotFound(problem);
-            }
-
-            if(userProjectService.CheckIfUserFollows(user.Id, projectId)== false)
-            {
-                ProblemDetails problem = new ProblemDetails
-                {
-                    Title = "User is not following this project",
-                    Detail = "You are not following this project.",
-                    Instance = "27D14082-9906-4EB8-AE4C-65BAEC0BB4FD"
-                };
-                return Conflict(problem);
-            }
-
-            Project project = await projectService.FindAsync(projectId);
-
-            if(await projectService.FindAsync(projectId) == null)
-            {
-                ProblemDetails problem = new ProblemDetails
-                {
-                    Title = "Failed getting the project.",
-                    Detail = "The database does not contain a project with this project id.",
-                    Instance = "57C13F73-6D22-41F3-AB05-0CCC1B3C8328"
-                };
-                return NotFound(problem);
-            }
-            UserProject userProject = new UserProject(project, user);
-            userProjectService.Remove(userProject);
-
-            userProjectService.Save();
-            return Ok();
-        }
+       
 
         /// <summary>
         ///     Delete the user account.
@@ -369,7 +266,7 @@ namespace API.Controllers
         /// </summary>
         /// <param name="followedUserId"></param>
         /// <returns></returns>
-        [HttpPost("follow/user/{followedUserId}")]
+        [HttpPost("follow/{followedUserId}")]
         [Authorize]
         public async Task<IActionResult> FollowUser(int followedUserId)
         {
@@ -423,7 +320,7 @@ namespace API.Controllers
         /// </summary>
         /// <param name="followedUserId"></param>
         /// <returns></returns>
-        [HttpDelete("follow/user/{followedUserId}")]
+        [HttpDelete("follow/{followedUserId}")]
         [Authorize]
         public async Task<IActionResult> UnfollowUser(int followedUserId)
         {
